+++ docs/examples/app/question_title02.py
@@ -15,7 +15,7 @@ class MyApp(App[str]):
         yield Button("No", id="no", variant="error")
 
     def on_button_pressed(self, event: Button.Pressed) -> None:
-        self.exit(event.button.id)
+        self.exit(event.sender.id)
 
     def on_key(self, event: Key):
         self.title = event.key